Add an option to enforce remaining in the camera limit
Context
=======
Godot Camera2D node has the attributes limit_left, limit_right,
limit_top and limit_bottom. Those attributes can be configured to make
sure that Godot won't move the camera out of it.
But if I have a tiny scene and if I reduce the zoom a lot, I may reach a
point where my whole scene becomes smaller than what the Camera2D
displays, which means that the Camera2D shows some "void".
(or, it's also possible to use those limit to display only a small part
 of the scene at first and to change those limit once the player
 performed some actions. In this case if the zoom becomes too small then
 the player can already see some of the not-yet-unlocked area)

What this commit does
=====================
This commit adds an option that prevent this behavior.

By default this new option is deactivated (so this commit keeps the
existing behavior). And when this option is activated it ensures
zoom the zoom cannot be smaller than this computed min_zoom.